agent_action/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-dotenv==1.0.0
# langchain-google-genai==0.0.5
# langchain-core==0.1.0
# langgraph==0.0.20
# pydantic==2.5.0
# PyMuPDF==1.23.8
# python-docx==1.1.0
# python-multipart
# grpcio
# grpcio-tools


fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain-google-genai==0.0.3
langchain-core==0.1.8
langgraph==0.0.10
pydantic==1.10.13
PyMuPDF==1.23.8
python-docx==1.1.0
python-multipart==0.0.9
grpcio==1.56.2
grpcio-tools==1.56.2
pymongo

